requirements.txt
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
python-decouple==3.8
requests==2.31.0
soupsieve==2.5
urllib3==2.1.0
